Rename summary.py to beam_summary.py
Pairs with shear_wall_summary.py, so the two summary modules are named
consistently.

mento.summary stays as a shim that re-exports BeamSummary and raises a
DeprecationWarning, since it is a published import path. The documented
entry point, `from mento import BeamSummary`, is unchanged.

Also fills in the API reference, which was missing entries for the
column, punching, shear_wall_summary and slab modules, and restores its
alphabetical order.
