+++ mento/__init__.py
@@ -92,7 +92,7 @@ if TYPE_CHECKING:
     from mento.settings import BeamSettings
     from mento.node import Node
     from mento.results import DocumentBuilder, Formatter, TablePrinter
-    from mento.summary import BeamSummary
+    from mento.beam_summary import BeamSummary
     from mento.column import Column
     from mento.punching import Capital, Opening, PunchingNode, PunchingSlab
     from mento.shear_wall import ShearWall
@@ -116,7 +116,7 @@ def __getattr__(name: str) -> object:
         "DocumentBuilder": "results",
         "EN_1992_2004_beam": "codes",
         "ACI_318_19_beam": "codes",
-        "BeamSummary": "summary",
+        "BeamSummary": "beam_summary",
         "Column": "column",
         "PunchingSlab": "punching",
         "Opening": "punching",
