+++ pyproject.toml
@@ -115,6 +115,7 @@ warn_return_any = false
 [[tool.mypy.overrides]]
 module = [
     "mento.beam",
+    "mento.beam_summary",
     "mento.codes.ACI_318_19_beam",
     "mento.codes.ACI_318_19_wall",
     "mento.codes.EN_1992_2004_beam",
@@ -124,7 +125,6 @@ module = [
     "mento.results",
     "mento.shear_wall",
     "mento.slab",
-    "mento.summary",
     "mento.units",
 ]
 ignore_errors = true
